requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.120
openai==1.45.0
postgrest==0.16.11
pypdf==4.3.1
python-dotenv==1.0.1
supabase==2.7.4